add short-circuit logic to HiGenParticleAna, also raise default minpt to 5

+++ HeavyIonsAnalysis/JetAnalysis/python/HiGenAnalyzer_cfi.py
@@ -3,7 +3,7 @@ import FWCore.ParameterSet.Config as cms
 HiGenParticleAna = cms.EDAnalyzer('HiGenAnalyzer',
 				  doVertex = cms.untracked.bool(False),
 				  etaMax = cms.untracked.double(2),
-				  ptMin = cms.untracked.double(0),
+				  ptMin = cms.untracked.double(5),
 				  chargedOnly = cms.untracked.bool(False),
 				  stableOnly = cms.untracked.bool(True),
 				  src = cms.untracked.InputTag("generator"),
